requirements.txt
asgiref==3.6.0
boto==2.49.0
Cactus==3.3.3
colorama==0.4.6
colorlog==6.7.0
Django==1.6.11
django-cors-headers==3.13.0
django-markwhat==1.6.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
keyring==5.7.1
markdown2==2.4.7
nodeenv==1.7.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
